Avoid storing the allocator in the task<T> coroutine frame if not needed.
In cases where the allocator_traits<_Alloc>::is_always_equal is true and
the allocator is default constructible then we avoid storing the allocator
and instead just construct a new one during deallocation.

Now also use allocator_traits<_Alloc>::rebind_alloc<char> to allocate the
coroutine frame instead of the allocator that was passed in.